feat: add security module
