+++ api/src/main/java/com/mm/api/item/dto/response/ItemResponse.java
@@ -1,11 +1,7 @@
-package com.mm.api.Item.dto.response;
+package com.mm.api.item.dto.response;
 
 import com.mm.coredomain.domain.Item;
 import com.mm.coredomain.domain.ItemCategoryType;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-
-import java.util.List;
 
 public record ItemResponse(Long id,
                            String detail,
