+++ core/core-security/src/main/java/com/example/coresecurity/CoreSecurityApplication.java
@@ -1,0 +1,13 @@
+package com.example.coresecurity;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CoreSecurityApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CoreSecurityApplication.class, args);
+	}
+
+}
