+++ core/core-security/src/test/java/com/example/coresecurity/CoreSecurityApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.coresecurity;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CoreSecurityApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
