+++ settings.gradle
@@ -3,3 +3,4 @@ rootProject.name = 'daldal'
 include 'api'
 include 'core:core-domain'
 include 'core:core-infra:core-infra-qdsl'
+include 'core:core-security'
